Refactoring added custom Controller components
